feat: future proof the code
make 2019 and all future years available.
this code can now run indefinetely because we only make assumptions
about earlier years.
Instead of adding a new year everytime we hit the first of January this
code should now work for all future years because we look at the current
year and from then backwards to 2017

update the code to use serverless-python requirements plugin to remove
the hack to incldue the virtualenv